Правки согласно комментариям ревью
