+++ ImageFeed/AppDelegate.swift
@@ -16,7 +16,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?
     ) -> Bool {
         ProgressHUD.animationType = .activityIndicator
-
         ProgressHUD.colorHUD = .black
         ProgressHUD.colorAnimation = .lightGray
