+++ ImageFeed/Models/Network/ProfileResult.swift
@@ -1,0 +1,14 @@
+//
+//  ProfileResult.swift
+//  ImageFeed
+//
+//  Created by Dmitriy Noise on 12.05.2025.
+//
+
+struct ProfileResult: Decodable {
+    let username: String
+    let firstName: String?
+    let lastName: String?
+    let bio: String?
+    
+}
