+++ ImageFeed/Models/Network/UrlsResult.swift
@@ -1,0 +1,12 @@
+//
+//  UrlsResult.swift
+//  ImageFeed
+//
+//  Created by Dmitriy Noise on 12.05.2025.
+//
+
+struct UrlsResult: Decodable {
+    let full: String
+    let small: String
+    let thumb: String
+}
