+++ ImageFeed/Models/Network/UserResult.swift
@@ -1,0 +1,10 @@
+//
+//  UserResult.swift
+//  ImageFeed
+//
+//  Created by Dmitriy Noise on 12.05.2025.
+//
+
+struct UserResult: Decodable {
+    let profileImage: [String:String]
+}
