+++ ImageFeed/Models/Profile.swift
@@ -1,0 +1,13 @@
+//
+//  Profile.swift
+//  ImageFeed
+//
+//  Created by Dmitriy Noise on 12.05.2025.
+//
+
+struct Profile {
+    let username: String
+    let name: String
+    let loginName: String
+    let bio: String
+}
