+++ ImageFeed/Service/ProfileService.swift
@@ -7,21 +7,6 @@
 
 import Foundation
 
-private struct ProfileResult: Codable {
-    let username: String
-    let firstName: String?
-    let lastName: String?
-    let bio: String?
-    
-}
-
-struct Profile {
-    let username: String
-    let name: String
-    let loginName: String
-    let bio: String
-}
-
 final class ProfileService {
     static let shared = ProfileService()
     private init() {}
